backend/requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
pandas>=1.3.3
numpy>=1.21.2
yfinance==0.2.36
pandas_datareader>=0.10.0
polygon-api-client>=0.2.11
python-dotenv>=0.19.0
pydantic>=1.8.2,<2.0.0
scikit-learn>=0.24.2